audio.pollinations.bak/requirements.txt
absl-py
accelerate
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
annotated-types
antlr4-python3-runtime
anyio
argbind
asttokens
attrs
audioread
blinker
boto3
botocore
certifi
cffi
charset-normalizer
click
contourpy
cycler
dacite
decorator
descript-audio-codec
descript-audiotools
docstring_parser
dotenv
einops
einx
executing
fastapi
ffmpy
filelock
fire
Flask
flask-cors
flatten-dict
fonttools
frozendict
frozenlist
fsspec
future
grpcio
h11
hf-xet
huggingface-hub
idna
importlib_resources
ipython
ipython_pygments_lexers
itsdangerous
jedi
jieba
Jinja2
jmespath
joblib
json_repair
julius
kiwisolver
langid
lazy_loader
librosa
llvmlite
loguru
Markdown
markdown-it-py
markdown2
MarkupSafe
matplotlib
matplotlib-inline
mdurl
more-itertools
mpmath
msgpack
multidict
networkx
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
omegaconf
openai-whisper
packaging
pandas
parso
pexpect
pillow
platformdirs
pooch
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pycparser
pydantic
pydantic_core
pydub
Pygments
pyloudnorm
pymon
pyparsing
pystoi
python-dateutil
python-dotenv
pytz
PyYAML
randomname
regex
requests
rich
ruff
s3fs
s3transfer
safetensors
scikit-learn
scipy
setuptools
six
sniffio
soundfile
soxr
stack-data
starlette
sympy
tensorboard
tensorboard-data-server
termcolor
threadpoolctl
tiktoken
tokenizers
torch
torch-stoi
torchaudio
torchvision
tqdm
traitlets
transformers
triton
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
vector-quantize-pytorch
wcwidth
Werkzeug
wrapt
yarl
gunicorn